refactor: simplify chat contact repo borrowing
